requirements.txt
llama-index-core>=0.10.0
llama-index-embeddings-google>=0.1.0
llama-index-embeddings-google-genai>=0.3.0
llama-index-llms-gemini>=0.1.0
llama-index-vector-stores-qdrant>=0.1.0
llama-index-graph-stores-neo4j>=0.1.0
llama-index-readers-file>=0.1.0
qdrant-client>=1.16.0
python-dotenv>=1.0.0
pymupdf>=1.23.0
neo4j>=5.0.0
